+++ samples/elaborate/contact_manager/contact_repository.rb
@@ -124,105 +124,8 @@ class ContactManager
       Jason
       Emma
       Olivia
-      Ava
-      Isabella
-      Sophia
-      Charlotte
-      Mia
-      Amelia
-      Harper
-      Evelyn
-      Abigail
-      Emily
-      Elizabeth
-      Mila
-      Ella
-      Avery
-      Sofia
-      Camila
-      Aria
-      Scarlett
-      Victoria
-      Madison
-      Luna
-      Grace
-      Chloe
-      Penelope
-      Layla
-      Riley
-      Zoey
-      Nora
-      Lily
-      Eleanor
-      Hannah
-      Lillian
-      Addison
-      Aubrey
-      Ellie
-      Stella
-      Natalie
-      Zoe
-      Leah
-      Hazel
-      Violet
-      Aurora
-      Savannah
-      Audrey
-      Brooklyn
-      Bella
-      Claire
-      Skylar
-      Lucy
-      Paisley
-      Everly
-      Anna
-      Caroline
-      Nova
-      Genesis
-      Emilia
-      Kennedy
-      Samantha
-      Maya
-      Willow
-      Kinsley
-      Naomi
-      Aaliyah
-      Elena
-      Sarah
-      Ariana
-      Allison
-      Gabriella
-      Alice
-      Madelyn
-      Cora
-      Ruby
-      Eva
-      Serenity
-      Autumn
-      Adeline
-      Hailey
-      Gianna
-      Valentina
-      Isla
-      Eliana
-      Quinn
-      Nevaeh
-      Ivy
-      Sadie
-      Piper
-      Lydia
-      Alexa
-      Josephine
-      Emery
-      Julia
-      Delilah
-      Arianna
-      Vivian
-      Kaylee
-      Sophie
-      Brielle
-      Madeline
     ]
+    
     NAMES_LAST = %w[
       Smith
       Johnson
@@ -235,8 +138,9 @@ class ContactManager
       Anderson
       Taylor
     ]
+    
     def initialize(contacts = nil)
-      @contacts = contacts || 1000.times.map do |n|
+      @contacts = contacts || 100.times.map do |n|
         random_first_name_index = (rand*NAMES_FIRST.size).to_i
         random_last_name_index = (rand*NAMES_LAST.size).to_i
         first_name = NAMES_FIRST[random_first_name_index]