+++ samples/hello/hello_combo.rb
@@ -23,30 +23,34 @@ class Person
   attr_accessor :country, :country_options
 
   def initialize
-    self.country_options=["", "Canada", "US", "Mexico"]
-    self.country = "Canada"
+    self.country_options = ['', 'Canada', 'US', 'Mexico']
+    reset_country
   end
 
   def reset_country
-    self.country = "Canada"
+    self.country = 'Canada'
   end
 end
 
 class HelloCombo
   include Glimmer
+  
   def launch
     person = Person.new
     
     shell {
-      fill_layout :vertical
-      text 'Hello, Combo!'
+      row_layout(:vertical) {
+        pack false
+      }        
+      
+      text 'Hello, Combo!'      
       
       combo(:read_only) {
         selection bind(person, :country)
       }
       
       button {
-        text "Reset Selection"
+        text 'Reset Selection'
         
         on_widget_selected do
           person.reset_country